+++ platformio/commands/test.py
@@ -35,7 +35,7 @@ from platformio.managers.platform import PlatformFactory
 @click.option("--project-dir", "-d", default=getcwd,
               type=click.Path(exists=True, file_okay=False, dir_okay=True,
                               writable=True, resolve_path=True))
-@click.option("--verbose", "-v", count=True, default=3)
+@click.option("--verbose", "-v", is_flag=True)
 @click.pass_context
 def cli(ctx, environment, skip, upload_port, project_dir, verbose):
     assert check_project_envs(project_dir, environment)