+++ platformio/managers/platform.py
@@ -23,7 +23,7 @@ from os.path import basename, dirname, isdir, isfile, join
 import click
 import semantic_version
 
-from platformio import exception, util
+from platformio import app, exception, util
 from platformio.managers.package import BasePkgManager, PackageManager
 
 PACKAGE_DIR = join(util.get_home_dir(), "packages")
@@ -224,11 +224,10 @@ class PlatformRunMixin(object):
         self.configure_default_packages(variables, targets)
         self.install_packages(silent=True)
 
-        self._verbose_level = int(verbose)
+        self._verbose = verbose or app.get_setting("force_verbose")
 
         if "clean" in targets:
-            targets.remove("clean")
-            targets.append("-c")
+            targets = ["-c", "."]
 
         variables['platform_manifest'] = self.manifest_path
 
@@ -237,14 +236,8 @@ class PlatformRunMixin(object):
         if not isfile(variables['build_script']):
             raise exception.BuildScriptNotFound(variables['build_script'])
 
-        self._found_error = False
         result = self._run_scons(variables, targets)
         assert "returncode" in result
-        # if self._found_error:
-        #     result['returncode'] = 1
-
-        if self._last_echo_line == ".":
-            click.echo("")
 
         return result
 
@@ -266,7 +259,10 @@ class PlatformRunMixin(object):
             "-j %d" % self.get_job_nums(),
             "--warn=no-no-parallel-support",
             "-f", join(util.get_source_dir(), "builder", "main.py")
-        ] + targets
+        ]
+        if not self._verbose and "-c" not in targets:
+            cmd.append("--silent")
+        cmd += targets
 
         # encode and append variables
         for key, value in variables.items():
@@ -280,31 +276,19 @@ class PlatformRunMixin(object):
         return result
 
     def on_run_out(self, line):
-        self._echo_line(line, level=3)
+        self._echo_line(line, level=1)
 
     def on_run_err(self, line):
         is_error = self.LINE_ERROR_RE.search(line) is not None
-        if is_error:
-            self._found_error = True
-        self._echo_line(line, level=1 if is_error else 2)
+        self._echo_line(line, level=3 if is_error else 2)
 
-    def _echo_line(self, line, level):
+    @staticmethod
+    def _echo_line(line, level):
         assert 1 <= level <= 3
-
-        fg = ("red", "yellow", None)[level - 1]
-        if level == 3 and "is up to date" in line:
+        fg = (None, "yellow", "red")[level - 1]
+        if level == 1 and "is up to date" in line:
             fg = "green"
-
-        if level > self._verbose_level:
-            click.secho(".", fg=fg, err=level < 3, nl=False)
-            self._last_echo_line = "."
-            return
-
-        if self._last_echo_line == ".":
-            click.echo("")
-        self._last_echo_line = line
-
-        click.secho(line, fg=fg, err=level < 3)
+        click.secho(line, fg=fg, err=level > 1)
 
     @staticmethod
     def get_job_nums():
@@ -326,13 +310,7 @@ class PlatformBase(PlatformPackagesMixin, PlatformRunMixin):
         self.pm = PackageManager(
             PACKAGE_DIR, self._manifest.get("packageRepositories"))
 
-        self._found_error = False
-        self._last_echo_line = None
-
-        # 1 = errors
-        # 2 = 1 + warnings
-        # 3 = 2 + others
-        self._verbose_level = 3
+        self._verbose = False
 
     @property
     def name(self):